% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_environment.R
\docType{class}
\name{volcanoExplorer}
\alias{volcanoExplorer}
\title{RL environment "volcano explorer", which is one of the mountain car problem.}
\format{An \code{\link{R6Class}} generator object.}
\usage{
volcanoExplorer
}
\description{
RL environment "volcano explorer", which is one of the mountain car problem.
}
\keyword{datasets}
